Created an API call with which a lecturer can request metadata about the lecture

+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/lectures/MockLectureServiceImplementation.java
@@ -22,7 +22,7 @@ public class MockLectureServiceImplementation implements LectureService {
     }
 
     @Override
-    public Map<String, Object> joinOngoingLecture(Long lectureId, Long joinId, String userName) {
+    public Map<String, Object> joinOngoingLecture(Integer lectureId, Long joinId, String userName) {
         Map<String, Object> toBeReturned = new TreeMap<>();
 
         toBeReturned.put("success", true);
@@ -34,7 +34,7 @@ public class MockLectureServiceImplementation implements LectureService {
     }
 
     @Override
-    public Map<String, Object> getLectureMetadata(Long lectureId) {
+    public Map<String, Object> getLectureMetadata(Integer lectureId) {
         Map<String, Object> toBeReturned = new TreeMap<>();
 
         toBeReturned.put("success", "true");
